Commenting out MD5 classes because it's not fully working yet

+++ src/main/java/homepage/algorithm/md5/Operation.java
@@ -1,8 +1,10 @@
+/*
 package homepage.algorithm.md5;
 
 @FunctionalInterface
 public interface Operation {
     
-    public int execute(int x, int y, int z);
+    public long execute(long x, long y, long z);
 
 }
+*/
